+++ BE/src/main/java/ifive/idrop/IDropApplication.java
@@ -7,9 +7,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @SpringBootApplication
 @EnableScheduling
 public class IDropApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(IDropApplication.class, args);
     }
-
 }
