+++ BE/src/main/java/ifive/idrop/child/domain/Child.java
@@ -1,6 +1,6 @@
 package ifive.idrop.child.domain;
 
-import ifive.idrop.pickup.domain.Subscription;
+import ifive.idrop.subscription.domain.Subscription;
 import ifive.idrop.common.enums.Gender;
 import ifive.idrop.parent.domain.Parent;
 import jakarta.persistence.*;
