+++ BE/src/main/java/ifive/idrop/notification/AlarmMessage.java
@@ -9,8 +9,7 @@ public enum AlarmMessage {
     PICK_UP_START("픽업 시작 안내", "기사가 픽업을 시작하였습니다"),
     PICK_UP_END("픽업 종료 안내", "기사가 픽업을 종료하였습니다"),
     SUBSCRIBE_REQUEST("구독 요청 안내", "구독 승인 요청이 왔습니다."),
-    PICK_UP_PRE_INFO("픽업 사전 안내", "픽업 시작까지 한 시간 남았습니다.")
-    ;
+    PICK_UP_PRE_INFO("픽업 사전 안내", "픽업 시작까지 한 시간 남았습니다.");
 
     private final String title;
     private final String message;
