--- BE/src/main/java/ifive/idrop/notification/NotificationRequest.java
@@ -1,14 +1,0 @@
-package ifive.idrop.notification;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class NotificationRequest {
-    private String title;
-    private String message;
-    private String token;
-}
